move daemon to separate package
for better testability
